rename connectivity data-transfer
