+++ extensions/data-transfer/portability-data-transfer-google/src/main/java/org/dataportabilityproject/transfer/google/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Implements importers and exporters for Google services.
  */
-package org.dataportabilityproject.connectivity.microsoft;
+package org.dataportabilityproject.transfer.google;
