+++ extensions/data-transfer/portability-data-transfer-microsoft/build.gradle
