+++ extensions/data-transfer/portability-data-transfer-microsoft/src/main/java/org/dataportabilityproject/transfer/microsoft/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Implements importers and exporters for Google services.
  */
-package org.dataportabilityproject.connectivity.google;
+package org.dataportabilityproject.transfer.microsoft;
